refactor the form into a component
